sql: Make INT an alias for INT8
This is the first in a series of changes to allow INT to eventually be
redefined as INT4.

Currently, a "naked" INT is generally treated as a 64-bit int type, but INT is
separate from INT8. This change eliminates INT as a distinct type in favor of
always using INT8. This is analogous to how we treat FLOAT/FLOAT8.

The coltypes.Int global has been removed and we should no longer produce type
objects that have a zero Width. The existing interpretation of zero-Width types
as 64-bit values will continue indefinitely to preserve backwards-compatibility
with existing descriptors.

In general, the changes to tests are due to the widespread use of INT as an
arbitary datatype. Except in a few cases where we are testing how "INT" is
interpreted as a datatype, the tests have been changed to simply use INT8, to
avoid another major round of text fixups as work on #26925 continues.

Release note (sql change): The INT type is now treated as an alias for INT8.

+++ pkg/cli/dump_test.go
@@ -376,7 +376,7 @@ func TestDumpAsOf(t *testing.T) {
 
 	const create = `
 	CREATE DATABASE d;
-	CREATE TABLE d.t (i int);
+	CREATE TABLE d.t (i int8);
 	INSERT INTO d.t VALUES (1);
 	SELECT now();
 `
@@ -397,7 +397,7 @@ func TestDumpAsOf(t *testing.T) {
 
 	const want1 = `dump d t
 CREATE TABLE t (
-	i INT NULL,
+	i INT8 NULL,
 	FAMILY "primary" (i, rowid)
 );
 
@@ -409,7 +409,7 @@ INSERT INTO t (i) VALUES
 	}
 
 	c.RunWithArgs([]string{"sql", "-e", `
-		ALTER TABLE d.t ADD COLUMN j int DEFAULT 2;
+		ALTER TABLE d.t ADD COLUMN j int8 DEFAULT 2;
 		INSERT INTO d.t VALUES (3, 4);
 	`})
 
@@ -419,8 +419,8 @@ INSERT INTO t (i) VALUES
 	}
 	const want2 = `dump d t
 CREATE TABLE t (
-	i INT NULL,
-	j INT NULL DEFAULT 2:::INT,
+	i INT8 NULL,
+	j INT8 NULL DEFAULT 2:::INT8,
 	FAMILY "primary" (i, rowid, j)
 );
